# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/otp/request": {
      "post": {
        "operationId": "AuthController_requestOtp",
        "parameters": [
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestResponseDto"
                }
              }
            }
          }
        },
        "summary": "Request a consumer OTP code by phone. No auth required.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/otp/verify": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumerAuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Verify a consumer OTP code and issue a token pair. No auth required.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/merchant/login": {
      "post": {
        "operationId": "AuthController_merchantLogin",
        "parameters": [
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantAuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Merchant email/password login. No auth required.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/admin/login": {
      "post": {
        "operationId": "AuthController_adminLogin",
        "parameters": [
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Admin email/password login. No auth required.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensDto"
                }
              }
            }
          }
        },
        "summary": "Rotate a refresh token for a fresh token pair. No auth required (the refresh token IS the credential).",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto"
                }
              }
            }
          }
        },
        "summary": "Revoke a refresh token's whole family. No auth required (the refresh token IS the credential).",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kurtar API",
    "description": "Kurtar — surplus-food marketplace API. The contract for all four\nfrontends (merchant-web, admin-web, the consumer app, and the\nlanding site).\n\n**Auth**: every route requires a bearer access token\n(`Authorization: Bearer <token>`) unless explicitly marked \"no auth\nrequired\" below — obtained via one of the /auth/* endpoints. A web\npanel additionally sends `X-Client-Transport: cookie` on every\nlogin/verify/refresh call to receive the refresh token ONLY as an\nhttpOnly cookie (never in the JSON body) — the mobile app omits\nthis header and reads the refresh token from the response body\ninstead.\n\n**Errors**: every error response is `{statusCode, errorCode,\nmessage}` (ErrorEnvelope, below) — branch on `errorCode`, never on\n`message` (not stable across releases, and may be Turkish or\nEnglish depending on the request).",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OtpRequestDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 20,
            "pattern": "PHONE_REGEX"
          }
        },
        "required": [
          "phone"
        ]
      },
      "OtpRequestResponseDto": {
        "type": "object",
        "properties": {
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "expiresAt"
        ]
      },
      "OtpVerifyDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 20,
            "pattern": "PHONE_REGEX"
          },
          "code": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": [
          "phone",
          "code"
        ]
      },
      "ConsumerAuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "BANNED",
              "DELETED"
            ]
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "phone",
          "status"
        ]
      },
      "ConsumerAuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string",
            "description": "Omitted when the caller declared cookie transport (X-Client-Transport: cookie) — the refresh token went out as an httpOnly cookie instead."
          },
          "refreshTokenExpiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "Omitted under the same condition as refreshToken."
          },
          "user": {
            "$ref": "#/components/schemas/ConsumerAuthUserDto"
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email"
          },
          "password": {
            "type": "string",
            "maxLength": 128
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "MerchantAuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "STAFF"
            ]
          },
          "merchantId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "merchantId"
        ]
      },
      "MerchantAuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string",
            "description": "Omitted when the caller declared cookie transport (X-Client-Transport: cookie) — the refresh token went out as an httpOnly cookie instead."
          },
          "refreshTokenExpiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "Omitted under the same condition as refreshToken."
          },
          "user": {
            "$ref": "#/components/schemas/MerchantAuthUserDto"
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
      "AdminAuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name"
        ]
      },
      "AdminAuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string",
            "description": "Omitted when the caller declared cookie transport (X-Client-Transport: cookie) — the refresh token went out as an httpOnly cookie instead."
          },
          "refreshTokenExpiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "Omitted under the same condition as refreshToken."
          },
          "user": {
            "$ref": "#/components/schemas/AdminAuthUserDto"
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
      "RefreshTokenBodyDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "maxLength": 256
          }
        }
      },
      "AuthTokensDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string",
            "description": "Omitted when the caller declared cookie transport (X-Client-Transport: cookie) — the refresh token went out as an httpOnly cookie instead."
          },
          "refreshTokenExpiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "Omitted under the same condition as refreshToken."
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "LogoutResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
...
    }
  }
}